move PostalCodeComparator to package facilities/common and add try/catch; correct postal code display format in file .ftlh
